pyproject.toml
[project]
name = "pp"
version = "0.1.0"
description = "Simple conversational backend server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "fastapi-cache2>=0.2.2",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.35.0",
    "opentelemetry-exporter-otlp-proto-http>=1.35.0",
    "opentelemetry-instrumentation-fastapi>=0.56b0",
    "opentelemetry-sdk>=1.35.0",
    "psycopg2-binary>=2.9.10",
    "py-bcrypt>=0.4",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "textual>=4.0.0",
    "textual-serve>=1.1.2",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "textual-dev>=1.7.0",
]


